requirements.txt
python-dotenv
networkx
natsort
